Refactor resolver and fix download blob bug for localfs mode
1. use common resolver for localfs mode and startgz
2. fix download blobs don't auth for localfs mode
